chore: move non-generated mocks from mock dir
