+++ services/wallet/market/feed_subscription_test.go
@@ -1,4 +1,4 @@
-package mock_common
+package market
 
 import (
 	"time"
