+++ services/wallet/market/market_feed_test.go
@@ -13,21 +13,19 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/event"
 
-	mock_common "github.com/status-im/status-go/services/wallet/common/mock"
-	mock_market "github.com/status-im/status-go/services/wallet/market/mock"
 	"github.com/status-im/status-go/services/wallet/thirdparty"
 )
 
 type MarketTestSuite struct {
 	suite.Suite
-	feedSub    *mock_common.FeedSubscription
+	feedSub    *FeedSubscription
 	tokensKeys []string
 	currencies []string
 }
 
 func (s *MarketTestSuite) SetupTest() {
 	feed := new(event.Feed)
-	s.feedSub = mock_common.NewFeedSubscription(feed)
+	s.feedSub = NewFeedSubscription(feed)
 
 	// Create test tokens
 	s.tokensKeys = []string{
@@ -46,7 +44,7 @@ func (s *MarketTestSuite) TestEventOnRpsError() {
 	defer ctrl.Finish()
 	// GIVEN
 	customErr := errors.New("request rate exceeded")
-	priceProviderWithError := mock_market.NewMockPriceProviderWithError(ctrl, customErr)
+	priceProviderWithError := NewMockPriceProviderWithError(ctrl, customErr)
 	manager := setupMarketManager(s.T(), []thirdparty.MarketDataProvider{priceProviderWithError}, s.feedSub.GetFeed())
 
 	// WHEN
@@ -65,7 +63,7 @@ func (s *MarketTestSuite) TestEventOnNetworkError() {
 
 	// GIVEN
 	customErr := errors.New("dial tcp: lookup optimism-goerli.infura.io: no such host")
-	priceProviderWithError := mock_market.NewMockPriceProviderWithError(ctrl, customErr)
+	priceProviderWithError := NewMockPriceProviderWithError(ctrl, customErr)
 	manager := setupMarketManager(s.T(), []thirdparty.MarketDataProvider{priceProviderWithError}, s.feedSub.GetFeed())
 
 	_, err := manager.FetchPrices(s.tokensKeys, s.currencies)
